HS-165 Add most rules for chord rules checker

+++ algorithms/src/main/scala/pl/agh/harmonytools/algorithm/graph/dijkstra/DijkstraNode.scala
@@ -1,6 +1,6 @@
 package pl.agh.harmonytools.algorithm.graph.dijkstra
 
 trait DijkstraNode {
-  def getDistanceFromBeginning: Int
+  def getDistanceFromBeginning: Double
   def getPrevsInShortestPath: List[DijkstraNode]
 }
